Moving mock list to mockData file
