+++ app/src/main/java/com/example/fitjournal/core/data/repository/RealmWorkoutLibraryRepositoryImpl.kt
@@ -18,96 +18,7 @@ class RealmWorkoutLibraryRepositoryImpl @Inject constructor() : RealmWorkoutLibr
         realm.write {
             val workouts: MutableList<RealmWorkoutLibrary> = mutableListOf()
             workouts.addAll(
-                listOf(
-                    MockData.libraryItem(
-                        "Bench",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Squats",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Lateral Raises",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Elevated Goblet Squats",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Burpees",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Bicep Curls",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Box Jumps",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Mountain Climbers",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Medicine Ball Slams",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Mountain Pose",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Modified Push-Ups",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Mason Twist",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Monkey Bars",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Modified Burpees",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Lunges",
-                        "Calisthenics"
-                    ),
-                    MockData.libraryItem(
-                        "Lat Pulldowns",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Lying leg curls",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "L-sit",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "LandMine Twists",
-                        "Weight Training"
-                    ),
-                    MockData.libraryItem(
-                        "Running",
-                        "Cardio"
-                    ),
-                    MockData.libraryItem(
-                        "Walking",
-                        "Cardio"
-                    ),
-                    MockData.libraryItem(
-                        "Biking",
-                        "Cardio"
-                    )
-                )
+                MockData.mockLibraryList
             )
             // running for each to simplify adding each individual workout entry
             workouts.forEach {
